==> code/3_plots_DEGanalysis.Rmd <==
---
title: "Plots_DEGanalysis (salmon)"
author: "Susanne Bauer"
date: "22/03/2021"
output: html_document
editor_options: 
  chunk_output_type: console

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```


```{r load packages}
library(tidyr)
library(ggplot2)
library(ggrepel)
library(ggforce)
library(dplyr)
library(plotly)
library(reshape2)
library(UpSetR)
library(kableExtra)
library(patchwork)
library(tximport)

date <- Sys.Date()

#make directory for output figures
dir <- paste0("./figures/DESeq2/")
ifelse(!dir.exists(dir), dir.create(dir),F) 

#make directory for results
file_dir <- paste0("./result_files/DESeq2/")
ifelse(!dir.exists(file_dir), dir.create(file_dir),F) #directory for output files


group_color_scheme <- c(FFI = "#B74D67", CJD = "#1D599C", Control = "#FEAF92")
celltype_color_scheme <- c(GadCer = "#50C1CA", PV = "#8B6EDA", SST = "#E0D42A", VglutCer = "#D55E00", GadCb = "#BDE8EB", VglutCb = "#EAAF80")
```

#Load DESeq2 results for all FFI and CJD cell types
```{r, concatenate DESeq2 genestat results}
#set path to results folder with DESeq2 output
path_FFI <-  "./result_files/intermediate_files/FFI/"
path_CJD <- "./result_files/intermediate_files/CJD/"

compile_genestats <- function(path){
  DEseq_list <- list.files(path, pattern = glob2rx("DESeq2*.csv"), full.names = T)
  
#extract TPM values and gene id from samples, sort by gene id and sum duplicates. Save TPM values with id in list
  genestats <- list()
  for (i in 1:length(DEseq_list)){
    name <- regmatches(DEseq_list[i], regexec('DESeq2_(.*?)\\.csv', DEseq_list[i]))[[1]][2]
    name <- strsplit(name, split = "_")
    Disease <- name[[1]][1]
    Celltype <- name [[1]][2]
    
    stat <- read.delim(DEseq_list[i], header=T, stringsAsFactors = F, sep = ",")
    stat$Celltype <- Celltype
    stat$Disease <- Disease
    genestats[[paste(Celltype)]] <- stat
    i <- i+1
  }
  
  all_stats <- bind_rows(genestats, .id = "Celltype")
  all_stats <- all_stats[complete.cases(all_stats),]
  all_stats <- all_stats %>%
    dplyr::select(Disease, Celltype, X, symbol, entrez, pvalue, padj, shrunkenLFC)%>%
    dplyr::rename(Ensembl = X, sLFC = shrunkenLFC, Entrez = entrez)
  return(all_stats)
}

CJD_genestats <- compile_genestats(path = path_CJD)
FFI_genestats <- compile_genestats(path = path_FFI)

all_genestats <- rbind(CJD_genestats, FFI_genestats)

#save compiled files
write.csv(CJD_genestats, file = paste0(file_dir, "CJD_DESeq_res_all.csv"))
write.csv(FFI_genestats, file = paste0(file_dir, "FFI_DESeq_res_all.csv"))
```

#Select only significant (FDR ≤ 0.05) DEGs
```{r format input data scatterplots}
#subset to keep only sig. DEGs
CJD_DEGs <- CJD_genestats %>% filter(padj < 0.05)
FFI_DEGs <- FFI_genestats %>% filter(padj < 0.05)

write.csv(CJD_DEGs, file = paste0(file_dir, "CJD_DEGs",date,".csv"))
write.csv(FFI_DEGs, file = paste0(file_dir, "FFI_DEGs",date,".csv"))

#set Celltypes as factors to determine order on plot
set_celltype_order <- function(data){
  data$Celltype <- factor(data$Celltype, 
                          levels = c("GadCer","PV","SST","VglutCer","GadCb", "VglutCb"))
  return(data)
}


#count number of DEGs by cell type, set count 0 if no DEGs detected in cell type
count_DEG <- function(data){
  data <- data %>% add_count(Celltype)
  data$n2 <- ifelse(is.na(data$sLFC), 0, data$n)
  return(data)
}

#create input data
FFI <- set_celltype_order(FFI_DEGs)
FFI <- count_DEG(FFI)

CJD <- set_celltype_order(CJD_DEGs)
CJD <- count_DEG(CJD)
```

#Jitter plots DEGs for all cell types by disease 
```{r scatter+violin plot showing DEGs by celltype}

#basic plot format
jitter_plot <- function(data){

  counts <- data %>%
    group_by(Celltype) %>%
    summarise(n = n())%>%
    filter(n >= 20)
  high_deg <- filter(data, Celltype %in% counts$Celltype)
  
  pos <- position_jitter(width = 0.2, height = 0.2)
  
  ggplot(data, mapping = aes(x = Celltype, y = sLFC, fill = padj, label = n2))+
      geom_jitter(position = pos, alpha = 0.5, size = 3, shape = 21, show.legend = T)+
      geom_text(data, mapping = aes(x = Celltype, y = n2), color = "grey52", y = Inf, vjust = 1, size = 8)+
      scale_fill_gradient2(mid = "#542788", high = "#FFC300")+
      geom_violin(high_deg, mapping = aes(x = Celltype, y = sLFC), alpha = 0.2, width = 1)+
      scale_x_discrete(labels = c(GadCer = "Gad2\nCerebrum", PV = "PV\nCerebrum", SST = "SST\nCerebrum", VglutCer = "vGluT2\nCerebrum", GadCb = "Gad2\nCerebellum", VglutCb = "vGluT2\nCerebellum"))+
      labs(fill = "FDR")+
      theme_bw() +
      theme(axis.title.x = element_blank(),
            axis.text.x = element_text(size = 16),
            axis.text.y = element_text(size = 14),
            legend.text = element_text(size = 14),
            panel.grid.major.y = element_line(colour = "grey95"),
            panel.grid.minor.y = element_blank(),
            panel.grid.major.x = element_blank(),
            panel.border = element_blank(),
            plot.subtitle = element_text(margin = margin(0,0,20,0)),
            plot.title.position = "panel",
            title = element_text(size = 20))
}

#CJD DEGs
set.seed(123)
CJD_jitter <- jitter_plot(CJD)
p1 <- CJD_jitter +
  ylim(-30,30)+
  labs(y = "log2 Fold Change",
       title = "Creutzfeldt-Jakob Disease (CJD)",
       subtitle = "Differentially expressed genes by cell type, FDR ≤ 0.05")
  #theme(legend.position = "none")
p1

#FFI DEGs
set.seed(123)
FFI_jitter <- jitter_plot(FFI)
p2 <- FFI_jitter +
  ylim(-30,30)+
  labs(y = "log2 Fold Change",
       title = "Fatal Familial Insomnia (FFI)",
       subtitle = "Differentially expressed genes by cell type, FDR ≤ 0.05")
p2

ggsave(plot = p1, file = paste0(dir, "CJD_DEGs_violin",date,".svg"), dpi = 300, units = "cm", width = 25, height = 15)

ggsave(plot = p2, file = paste0(dir, "FFI_DEGs_violin",date,".svg"), dpi = 300, units = "cm", width = 25, height = 15)


#Jitter plots with zoom_in -> outliers not removed from data
CJD_zoom <- p1 
CJD_zoom$layers[[3]] <- NULL #remove violin plots from zoom-in
CJD_zoom <- CJD_zoom + coord_cartesian(ylim = c(-13,10))
CJD_zoom
ggsave(plot = CJD_zoom, file = paste0(dir, "CJD_DEGs_violin_zoom",date,".svg"), dpi = 300, units = "cm", width = 25, height = 15)


FFI_zoom <- p2 
FFI_zoom$layers[[3]] <- NULL
FFI_zoom <- FFI_zoom + coord_cartesian(ylim = c(-13,10))
FFI_zoom
ggsave(plot = FFI_zoom, file = paste0(dir, "FFI_DEGs_violin_zoom",date,".svg"), dpi = 300, units = "cm", width = 25, height = 15)
```

#Zoom-in jitter plot for SST neurons, highlighting ribosomal proteins
```{r}
#zoom_in SST
prion <- rbind(FFI, CJD)
set.seed(123)
jitter_zoom <- function(data){
  rpp <- data %>%
    filter(grepl("Rp", symbol))%>%
    pull(symbol)
  
  data <- data %>%
    filter(Celltype == "SST")%>%
    mutate("RP" = case_when(symbol %in% rpp ~T,
                            T ~F))
  
  pos <- position_jitter(width = 0.2, height = 0.2)
  
  ggplot(data, mapping = aes(x = Celltype, y = sLFC, label = n2, fill = RP))+
      #geom_violin(alpha = 0.1, width = 1)+
      geom_point(aes(alpha = RP), position = pos, size = 2.5, shape = 21)+
      #geom_text(data, mapping = aes(x = Celltype, y = n2), color = "grey52", y = -27, size = 8)+
      #geom_text_repel(aes(label = ifelse(symbol %in% rpp, symbol, "")), position = pos, max.overlaps = Inf, min.segment.length = 0)+
      scale_fill_manual(label = c("DEG", "Ribosomal protein"), values = c("#D1E7F3","#51ABDD"))+
      guides(alpha = F)+
      ylim(-2,2)+
      facet_grid(~Disease)+
      theme_minimal() +
      theme(axis.title.x = element_blank(),
            axis.text.x = element_blank(),
            axis.text.y = element_text(size = 12),
            legend.text = element_text(size = 12),
            panel.grid.major.y = element_line(colour = "grey95"),
            panel.grid.minor.y = element_blank(),
            panel.grid.major.x = element_blank(),
            panel.border = element_blank(),
            plot.subtitle = element_text(margin = margin(0,0,20,0)),
            plot.title.position = "panel",
            strip.text = element_text(size = 16),
            legend.position = "bottom")+
    labs(y = "log2 Fold Change",
         x = "",
         fill = "")
}

dotplot_RPs <- jitter_zoom(prion)
dotplot_RPs

ggsave(file = paste0(dir, "both_violin_zoom_RPs",date,".svg"), dpi = 300, units = "cm", width = 14, height = 8)
```


#Barplot of number of up/downregulated and shared DEGs between diseases
```{r}
Prion <- rbind(FFI, CJD)

get_shared_DEG <- function(df){
  ct <- levels(Prion$Celltype)
  l <- list()

  for (i in 1:length(ct)) {
    z <- ct[i]
    p <- df[complete.cases(df),]
    p <- p %>% filter(Celltype == z)
    q <- as.vector(p %>% 
      filter(Disease == "CJD") %>%
      dplyr::select(symbol)) %>%
      pull(symbol)
    r <- as.vector(p %>% 
      filter(Disease == "FFI") %>%
      dplyr::select(symbol)) %>%
      pull(symbol)
    s <- intersect(q,r)
    
    t <- Prion %>%
    filter(Celltype == z)%>%
    mutate(shared_DEG = case_when(
      symbol %in% s ~ "shared",
      TRUE ~ "unique"))
  
  l[[paste(z)]] <- t
      
  i <- i+1
  }
  
  return(l)
}

p <- get_shared_DEG(Prion)
Prion <- do.call(rbind, p)

mat <- Prion %>%
  mutate(Direction = sign(sLFC)) %>%
  group_by(Celltype, Disease, shared_DEG)  %>%
  dplyr::count(Direction) %>%
  mutate(Dir = case_when(Direction > 0 ~ "up",
                         Direction < 0 ~ "down"))%>%
  mutate(n2 = n*Direction)%>%
  mutate(shared = paste(shared_DEG, Disease))%>%
  mutate(shared = case_when(shared == "shared FFI" ~"shared",
                            shared == "shared CJD" ~"shared",
                            TRUE ~shared))
mat$Celltype <- factor(mat$Celltype, labels = c(GadCer = "Gad2\nCerebrum", PV = "PV\nCerebrum", SST = "SST\nCerebrum", VglutCer = "vGluT2\nCerebrum", GadCb = "Gad2\nCerebellum", VglutCb = "vGluT2\nCerebellum"))

mat_nSST <- mat %>% filter(Celltype != "SST\nCerebrum")
mat_SST <- mat %>% filter(Celltype == "SST\nCerebrum")

make_DEG_barplot <- function(data){
  plot <- ggplot(data, aes(x = Disease, y = n2, fill = shared, label = n))+
  geom_col(position = position_stack(reverse = T))+
  geom_hline(yintercept = 0, color = "white")+
  scale_fill_manual(values = c("#FEAF92", "#1D599C", "#B74D67"))+
  scale_y_continuous(labels = abs)+
  theme_minimal()+
  theme(
    panel.grid.minor = element_blank(),
    panel.grid.major.x = element_blank(),
    axis.text.x = element_blank(),
    axis.text.y = element_text(size = 14),
    plot.title = element_text(size = 20),
    legend.text = element_text(size = 14),
    legend.title = element_text(size = 14)
  )+
    labs(y = "",
       x = "",
       fill = "FDR ≤ 0.05")+
  facet_grid(~Celltype,  drop = T, scales = "free_y", space = "free_y", shrink = F,  margin = T)+
  theme(strip.text.x = element_text(size = 12))
  
  return(plot)
}

bar_nSST <- make_DEG_barplot(mat_nSST)
bar_nSST
ggsave(filename = paste0(dir, "bar_sharedDEG_nSST.svg"), units = "cm", width = 18, height = 10)

bar_SST <- make_DEG_barplot(mat_SST)
bar_SST + theme(legend.position = "none", plot.title = element_blank())
ggsave(filename = paste0(dir, "bar_sharedDEG_SST.svg"), units = "cm", width = 4, height = 10)
  
```


#Upset plots of shared DEGs between cell types for each disease
```{r Upset plots showing shared genes}
#Upset plot showing number of shared genes between cell types and diseases
get_upset_input_disease <- function(df){
  require(dplyr)
    
  mat <- df[complete.cases(df),] #remove any NAs as they would be counted as shared genes
  mat <- dplyr::select(mat, "Celltype", "symbol") 
  mat <- split(mat, f = mat$Celltype) #splits df by cell type, return list of lists by cell type
  mat <- list(lapply(mat, "[",, "symbol")) #selects list with gene names from each cell type list
  mat <- mat[[1]]
  return(mat)
}

upset_FFI <- get_upset_input_disease(FFI)
upset_CJD <- get_upset_input_disease(CJD)

#make upset plot by celltype
plot_FFI <- upset(fromList(upset_FFI), 
                   nsets = 6, 
                   point.size = 5, 
                   line.size = 2, 
                   text.scale = c(3, 2, 3, 2, 3, 3),
                   mb.ratio = c(0.5, 0.5),
                   sets.x.label = "DEGs (FFI)")
plot_FFI
png(filename = paste0(dir, "upset_FFI",date,".svg"), height = 15, width = 25, units = "cm", res = 300)
plot_FFI
dev.off()

#while (!is.null(dev.list()))  dev.off()

plot_CJD <- upset(fromList(upset_CJD), 
                   nsets = 6, 
                   point.size = 5, 
                   line.size = 2, 
                   text.scale = c(3, 2, 3, 2, 3, 3),
                   mb.ratio = c(0.5, 0.5),
                   sets.x.label = "DEGs (CJD)"
                   ) 
png(filename = paste0(dir, "upset_CJD",date,".svg"), height = 15, width = 20, units = "cm", res = 300)
plot_CJD
dev.off()
plot_CJD
```

#Extract name of shared genes between cell types for each disease
```{r}
#function extracts list of intersection members from Upset plot. Unlike plot, which doesn't allow duplicates, it gives all members of the intersection. Uses same input list as "upset" function (before "fromList" transformation)

get_intersect_members <- function(c){
  
  #makes list with all possible combinations of elements in input list -> all possible cell type intersections
 
  combos <- Reduce(c,lapply(2:length(c), 
            function(x) combn(1:length(c),x,simplify=FALSE) ))

  #get intersecting elements from all possible vector combos (here: shared genes between cell types)
  intersect <- lapply(combos, function(x) Reduce(intersect,c[x]) )

  #make list names readable by replacing number with label specifying compared cell type
  levels <- c(1:6)
  labels <- data.frame(lapply(combos, "[", 1:6))
  labs <- c(names(c))
  for (i in 1:length(levels)){
    labels[labels == i] <- labs[i]
    i <- i+1
  }
  
  #combines cell types in intersection into single string to use as name
  new_labels <- list()
  for (i in 1:length(labels)){
    v <- labels[,i][!is.na(labels[,i])]
    v <- toString(v)
    new_labels <- append(new_labels, v)
  }
  names(intersect) <- lapply(new_labels, "[", 1)
  return(intersect)
}

CJD_intersect <- get_intersect_members(upset_CJD)
FFI_intersect <- get_intersect_members(upset_FFI)

#transforms list with intersections into data frame
intersect_table <- function(input){
  max_len <- max(lengths(input)) #get length of longest intersection
  df <- data.frame(t(do.call(cbind.data.frame, c(lapply(input, function(x) 
               c(x, rep("", max_len - length(x)))), stringsAsFactors = FALSE)))) #combine list elements in dataframe with length max_len
  df <- df %>% unite(Shared_genes, everything(), sep = "/ ", remove = T, na.rm = T) #collapse all gene names into one column, separated by ","
  df$count <- as.vector(lengths(input)) #add column with count of genes in intersection
  df <- df[df$count != 0,] #remove empty intersections
  #create output table 
  table <- kable(df, format = "html", row.names = T) %>%
  kable_styling(full_width = T) %>%
   column_spec(column = 1, bold = T)
  
  return(table)
}

FFI_table <- intersect_table(FFI_intersect)
FFI_table

CJD_table <- intersect_table(CJD_intersect)
CJD_table
```

#Compile excel output file of DEGs including TPM values from IPs
```{r}
metadata <- read.csv("./metadata/prionsamples_meta.csv", header = T, sep = ";")

#subset which samples to analyze
samples_meta <- metadata %>%
  filter(sample_set != "exclude")

#list salmon output files (quant.sf) for selected samples
files <- file.path("./salmon/",samples_meta$nfcore_id, "quant.sf")

#assign sample id to file names to rename samples based on id in txi output
names(files) <- samples_meta$sample_id  

tx2gene <- read.table("./metadata/salmon_tx2gene.tsv")

#read in salmon output files 
txi <- tximport(files, type = "salmon", tx2gene = tx2gene)
#tximport generates several output files. "count" matrix containes raw counts summarized to gene level, which are used for DESeq2 analysis. "abundance" gives tpm values summarized to gene level.

#check that sample order in count and metadata files match
sample_ids <- samples_meta$sample_id
all(colnames(txi$counts) == sample_ids)

all_TPMs <- txi$abundance

#extract and subset ERCC counts
ERCC_idx <- grep(pattern = "ERCC*", rownames(all_TPMs))
ERCC_TPMs <- data.frame(all_TPMs[ERCC_idx,])
all_TPMs <- data.frame(all_TPMs[-ERCC_idx,])

rownames(all_TPMs) <- stringr::str_replace(rownames(all_TPMs), pattern = ".[0-9]+$", replacement = "")

#########CHANGE HERE############
library(stringr)

#Filter IP samples and adjust cell type nomenclature
samples_meta <- samples_meta %>%
  filter(Fraction == "IP")%>%
  filter(sample_set != "exclude") %>%
  mutate(int = interaction(Cell_type, Region))%>%
  mutate(int = case_when(int == "Gad2.Cerebellum" ~ "GadCb",
                         int == "Gad2.Cerebrum" ~ "GadCer",
                         int == "PV.Cerebrum" ~ "PV",
                         int == "SST.Cerebrum" ~ "SST",
                         int == "vGlut2.Cerebellum" ~ "VglutCb",
                         int == "vGlut2.Cerebrum" ~ "VglutCer"))



#get select TPM values for samples  specific cell types
library(xlsx)
ct <- as.character(unique(samples_meta$int))

df_DEG_w_TPM <- list()

for (i in 1:length(ct)) {
  v <- ct[i]
  smpl <- samples_meta %>%
    filter(int == v) %>%
    pull(sample_id)
  
  DEGs <- Prion %>%
    filter(Celltype == v)
  DEG_ens <- unique(DEGs$Ensembl)

  tpm <- all_TPMs %>%
    filter(rownames(.) %in% DEG_ens)%>%
    dplyr::select(all_of(smpl))
 DEG_tpm <- merge(DEGs, tpm, by.x = "Ensembl", by.y = 0)
 
  ##get vector of genes with zero values in Ctrl and disease group
  cjd_s <- samples_meta %>% filter(int == v & Group == "CJD") %>% pull(sample_id)
  ffi_s <- samples_meta %>% filter(int == v & Group == "FFI") %>% pull(sample_id)
  ctrl_s <- samples_meta %>% filter(int == v & Group == "Control") %>% pull(sample_id)
  
  zero_counts <- c()
  for (n in 1:length(DEG_tpm$Ensembl)){
       g <- DEG_tpm[n,] 
       cjd_z <- sum(g[,cjd_s] == 0, na.rm = T) #return logical vector, TRUE if tpm == 0, count TRUEs, if >0 group contains zero sample
       ffi_z <- sum(g[,ffi_s] == 0, na.rm = T)
       ctrl_z <- sum(g[,ctrl_s] == 0, na.rm = T)
       
       zero <- case_when(g$Disease == "CJD" & cjd_z>0 & ctrl_z>0 ~T, #print true if both disease and ctrl group contain zero values
                        g$Disease == "FFI" & ffi_z>0 & ctrl_z>0 ~T,
                        TRUE ~F)
       
       zero_counts[n] <- zero
       n <- n+1
   } 
 DEG_tpm$zero_counts <- zero_counts #include vector with zero_ct in file
 df_DEG_w_TPM[[paste(v)]] <- DEG_tpm %>% select(Disease, Celltype, symbol, sLFC, padj, pvalue, shared_DEG, zero_counts, n, n2)
 
 DEG_tpm <- DEG_tpm %>%
   dplyr::select(-n2, -n)
 
 write.xlsx(DEG_tpm, file = paste0(file_dir, "DEG_with_TPM_zero",date,".xlsx"), sheetName = v, append = T)
 
 i <- i+1
}

DEG_zero_rm <- do.call("rbind", df_DEG_w_TPM)

```
